# services/shipping-service/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/shipments": {
            "get": {
                "description": "Get shipments with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Get shipments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new shipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Create shipment",
                "parameters": [
                    {
                        "description": "Shipment Create Request",
                        "name": "shipment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.ShipmentCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shipments/order/{order_id}": {
            "get": {
                "description": "Get a shipment by order ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Get shipment by order ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shipments/status/{status}": {
            "get": {
                "description": "Get shipments by status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Get shipments by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shipments/tracking/{tracking_number}": {
            "get": {
                "description": "Get a shipment by tracking number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Get shipment by tracking number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tracking Number",
                        "name": "tracking_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shipments/{id}": {
            "get": {
                "description": "Get a shipment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Get shipment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a shipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Update shipment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Shipment Update Request",
                        "name": "shipment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.ShipmentUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a shipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Delete shipment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shipments/{id}/status": {
            "patch": {
                "description": "Update a shipment status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipments"
                ],
                "summary": "Update shipment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.Shipment": {
            "type": "object",
            "required": [
                "order_id",
                "tracking_number"
            ],
            "properties": {
                "actual_delivery": {
                    "type": "string"
                },
                "carrier": {
                    "type": "string",
                    "maxLength": 100
                },
                "created_at": {
                    "type": "string"
                },
                "estimated_delivery": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "shipping_method": {
                    "type": "string",
                    "maxLength": 50
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "in_transit",
                        "delivered",
                        "returned"
                    ],
                    "example": "pending"
                },
                "tracking_number": {
                    "type": "string",
                    "maxLength": 100
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.ShipmentCreateRequest": {
            "type": "object",
            "required": [
                "order_id"
            ],
            "properties": {
                "carrier": {
                    "type": "string",
                    "maxLength": 100
                },
                "estimated_delivery": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "shipping_method": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "github_com_DurgaPratapRajbhar_e-commerce_shipping-service_internal_domain_entity.ShipmentUpdateRequest": {
            "type": "object",
            "properties": {
                "actual_delivery": {
                    "type": "string"
                },
                "carrier": {
                    "type": "string",
                    "maxLength": 100
                },
                "estimated_delivery": {
                    "type": "string"
                },
                "shipping_method": {
                    "type": "string",
                    "maxLength": 50
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "in_transit",
                        "delivered",
                        "returned"
                    ]
                },
                "tracking_number": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
    }
}